# perf: [Geneva Exporter] Optimize encoding