## Changes
- Call `create_schema` only when needed
- Avoid unnecessary allocation of `Vec<FieldDef>` in `create_schema` method

## Benchmarks

Benchmarks for encoding show **more than 70% perf improvement** for batch sizes of 100 and above

Machine details:

OS: Ubuntu 22.04.4 LTS (5.15.153.1-microsoft-standard-WSL2)
Hardware: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz, 16vCPUs,
RAM: 64.0 GB

```
encode_log_batch_sizes/batch_size/1
  main: time:   [6.9646 µs 6.9799 µs 6.9984 µs]
  PR:   time:   [6.7596 µs 6.7881 µs 6.8247 µs]  −3.0543%

encode_log_batch_sizes/batch_size/10
  main: time:   [40.525 µs 40.590 µs 40.659 µs]
  PR:   time:   [15.697 µs 15.762 µs 15.837 µs]  −60.983%

encode_log_batch_sizes/batch_size/100
  main: time:   [380.36 µs 381.00 µs 381.73 µs]
  PR:   time:   [108.90 µs 109.26 µs 109.64 µs]  −70.794%

encode_log_batch_sizes/batch_size/1000
  main: time:   [3.8238 ms 3.8734 ms 3.9402 ms]
  PR:   time:   [1.0539 ms 1.0568 ms 1.0601 ms]  −72.495%
```